(Maybe) added support for cabinets field
To enable that feature, uncomment lines 543, 555, 608 and 628
Also check if 'dataType': 'json' works on line 644
ALSO we should remove `|| true` on line 179 at some point to enable
caching teacher loads (see line 178 for more info).

BTW I think that will be a LOT better if I just rewrite that entire
page, because there is a lot of *hacks* in CSS, schedule displaying is
VERY slow and MUST be rewritten eventually.